Squash feature/add-text-templates (#54)
* Added static strings, text templates and table templates
* Removed colors command
* Added map rot generation on map rot switch
* Fixed automatic map remove

+++ src/wesenGemaMod/Map/MapRemover.lua
@@ -7,6 +7,7 @@
 
 local Exception = require("Util/Exception");
 local MapHandler = require("Map/MapHandler");
+local TextTemplate = require("Output/Template/TextTemplate");
 
 ---
 -- Handles removing of maps.
@@ -46,25 +47,30 @@ getmetatable(MapRemover).__call = MapRemover.__construct;
 function MapRemover:removeMap(_dataBase, _mapName, _mapRot)
 
   local mapId = MapHandler:fetchMapId(_dataBase, _mapName);
+  if (mapId) then
 
-  if (self:mapHasRecords(_dataBase, mapId)) then
-    error(Exception("Could not delete map \"" .. _mapName .. "\": There are map records for this map!"));
-  else
+    if (self:mapHasRecords(_dataBase, mapId)) then
+      error(Exception(
+          TextTemplate(
+            "ExceptionMessages/MapRemover/MapRecordsExistForDeleteMap",
+            { ["mapName"] = _mapName }
+          )
+      ));
+    else
 
-    -- Remove the map from the database
-    local sql = "DELETE FROM maps "
-             .. "WHERE id=" .. mapId .. ";";
-    _dataBase:query(sql, false);
-
-    -- Remove the map from the map rot
-    _mapRot:removeMap(_mapName);
-
-    -- Remove the map cgz and cfg files
-    removemap(_mapName);
-
-    return true;
+      -- Remove the map from the database
+      local sql = "DELETE FROM maps "
+               .. "WHERE id=" .. mapId .. ";";
+      _dataBase:query(sql, false);
+    end
 
   end
+
+  -- Remove the map from the map rot
+  _mapRot:removeMap(_mapName);
+
+  -- Remove the map cgz and cfg files
+  removemap(_mapName);
 
 end
 
@@ -80,7 +86,7 @@ function MapRemover:mapHasRecords(_dataBase, _mapId)
 
   local sql = "SELECT id FROM records "
            .. "WHERE map=" .. _mapId
-           .. "LIMIT 1;";
+           .. " LIMIT 1;";
   local result = _dataBase:query(sql, true);
 
   if (#result == 0) then
